put evaluator into client package
