+++ client/client.controller.js
@@ -1,24 +1,26 @@
-/**
- * To be mounted at / (root).
- */
-const express = require('express');
-const router = express.Router();
+'use strict';
+
 // Utils
 const thenable = require('@splitsoftware/splitio/lib/utils/promise/thenable');
 const reduce = require('lodash/reduce');
 const map = require('lodash/map');
 const config = require('config');
+
 // Own modules
 const utils = require('../utils');
+const common = require('./common');
 const sdkModule = require('../sdk');
+
 // Client and manager we will use
 const client = sdkModule.client;
 const manager = sdkModule.manager;
 
 /**
- * get-treatment endpoint. Evaluates a given split.
+ * getTreatment evaluates a given split-name
+ * @param {*} req 
+ * @param {*} res 
  */
-router.get('/get-treatment', (req, res) => {
+const getTreatment = (req, res) => {
   console.log('Getting a treatment.');  
   const state = req.query;
   const key = utils.parseKey(state.key, state['bucketing-key']);
@@ -47,12 +49,14 @@ router.get('/get-treatment', (req, res) => {
 
   if (thenable(eventuallyAvailableValue)) eventuallyAvailableValue.then(asyncResult);
   else asyncResult(eventuallyAvailableValue);
-});
+};
+
 /**
- * get-treatments endpoint. 
- * Returns the evaluations for all treatments matching the traffic type of the provided keys.
+ * getTreatments  returns the evaluations for all treatments matching the traffic type of the provided keys.
+ * @param {*} req 
+ * @param {*} res 
  */
-router.get('/get-treatments', (req, res) => {
+const getTreatments = (req, res) => {
   console.log('Getting treatments.');    
   const state = req.query;
   let keys = [];
@@ -80,7 +84,7 @@ router.get('/get-treatments', (req, res) => {
       return {
         trafficType: key.trafficType,
         key: utils.parseKey(key.matchingKey, key.bucketingKey),
-        splits: filterSplitsByTT(views, key.trafficType)
+        splits: common.filterSplitsByTT(views, key.trafficType)
       };
     });
   });
@@ -109,20 +113,9 @@ router.get('/get-treatments', (req, res) => {
     })
     // 500 on error
     .catch(() => res.sendStatus(500));
-});
+};
 
-/**
- * Utility function. Reduces a collection of split views to a list of names of the splits
- * corresponding to the given traffic type.
- */
-function filterSplitsByTT(splitViews, trafficType) {
-  return reduce(splitViews, (acc, view) => {
-    if (view.trafficType === trafficType) {
-      acc.push(view.name);
-    }
-    return acc;
-  }, []);
-}
-
-module.exports = router;
-
+module.exports = {
+  getTreatment,
+  getTreatments,
+};
